requirements.txt
Flask==2.0.2
scikit-learn==0.24.2
scikit-misc==0.1.4
librosa==0.8.1
numpy==1.21.0
torchaudio==2.0.2
torchsummary==1.5.1